requirements-api.txt
# Core ML dependencies
transformers
huggingface_hub
torch
pillow
python-dotenv

# API dependencies
fastapi
uvicorn[standard]
python-multipart

# Optional: for better performance
# gunicorn  # For production deployment on Linux
